Save, Load
Save, Load system that currently saves the player's last location as well as the health at Save.
Also inclusion of a collectible system for testing but not yet implemented.

Includes the removal of the old Player and movement and inclusion of the newer one.
(Currently Play Game takes to SaveandSaveAndSaveAndSaver for testing for Load functionality).

Next: Load from Menu button,
Saving and Loading State of Enemies