SaveLoadSystem
Additions to Checkpoint System, GameManager, PlayerSave, SaveLoadSystem as well as MainMenu scripts.
Load from the Main Menu now functional and loads in player with the variables from the last session.

Script added to allow for Checkpoint Menu to pop up, along with Save button but after testing, is not functional.

Additionally Colliders for Checkpoints were adjusted, checkpoint Menu assigned in the inspector as well as the Player object, to th script for the Checkpoints.

LoadGame function included in the existing MainMenu script to allow for the Load functions to be called when the button is pressed and the variables attached to player after.