Delivery Drone Prefab/Javelin Prefab Mesh/ Javelin VFX Update
Full prefab set up for the delivery drone prefab (the enemy that we utilize as a dash to target). The level needs to swap every instance of the package drone copy to this prefab in order to implement this change. Javelin now uses the mesh design and the vfx associated with it. Javelin explosion effects have been modified to Peter's visual effect. The trail effect on the javelin was update with a VFXcontroller script to disable during aiming (preventing visual clutter) and enabling the trail on launch (showing a trail behind the javelin as it soars).